pyproject.toml
[project]
name = "poc-crewai"
version = "0.1.0"
description = "Proof of concept for crewAI"
authors = [{ name = "User", email = "user@example.com" }]
dependencies = [
    "crewai>=0.28.8",
    "crewai[tools]>=0.1.6",
    "python-dotenv>=1.0.0",
    "langchain-groq>=0.1.3",
    "langchain-openai>=0.1.0", # For OpenAI ChatGPT
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.build.targets.wheel]
packages = ["poc_crewai"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv]
dev-dependencies = []
